feat: 取引先機能の型定義とAPIフックの整理
- 新しい型定義ファイル「businessPartner.ts」を作成し、取引先企業関連の型を明確化
- 取引先詳細取得のカスタムフック「useBusinessPartnerDetail」を追加し、APIからのデータ取得を簡素化
- 取引先一覧取得のカスタムフック「useBusinessPartners」を導入し、API呼び出しのロジックを整理
- 取引先詳細ページと一覧ページでカスタムフックを使用し、コードの可読性を向上

これにより、取引先機能の実装がより明確になり、メンテナンス性が向上します。
