+++ CLAUDE.md
@@ -9,7 +9,9 @@
 1. **Serenaとの連携**
    - 必ずSerenaと協力して作業を進めること。
 
-2. **事前理解**
+2. **前提**
+
+2.1 **事前理解**
    - 作業前に必ず`CompactSkillSheetsApp/_Documents`ディレクトリ内の全ファイルを理解すること。
    - 行動前に`CompactSkillSheetsApp/_Knowledge`ディレクトリも読み込むこと。
    - `CompactSkillSheetsApp/README.md`の環境構造を理解すること。
@@ -20,8 +22,12 @@
    - **必須**既存機能を壊さないこと
    - **必須**開発環境DockerのPort設定は壊さないこと
    - **必須**必要な実装か不要な実装か、全体を見て正しい判断をしてから実装すること
+
+2.2. **禁止事項**
+   - **必須**暫定対応の禁止
    - **必須**ハードコーディングは絶対に使用禁止
    - **必須**ダミーデータを画面に表示するのは絶対に禁止（DBの値を参照して表示すること）
+
 
 3. **実装ポリシー**
    - Kent Beck氏のテスト駆動開発（TDD）を実践。
